Small comment

+++ src/Readability.php
@@ -260,7 +260,8 @@ class Readability extends Element implements ReadabilityInterface
         if (get_class($this->node) !== 'DOMDocument') {
 
             // To prevent the -0 value
-            $this->contentScore = ($score === (float) -0) ? 0 : $score;
+            // TODO Might not be necessary now
+            $this->contentScore = ($score === (double)-0) ? 0 : $score;
 
             // Set score in an attribute of the tag to prevent losing it while creating new Readability objects.
             $this->node->setAttribute('readability', $this->contentScore);
